[Refactor] #113 이미지로 렌더링 해주는 함수 로직 개선
- extension UIImage를 UITableView로 변환
- 함수 이름 변경
- 사용하지 않는 로직 제거
- 강제 언래핑 개선
